% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfEnv.R
\name{dfEnv}
\alias{dfEnv}
\title{Create R Environment from data frame}
\usage{
dfEnv(df, orientation = "cols", global = TRUE)
}
\arguments{
\item{df}{A \code{data frame}.}

\item{orientation}{One of \code{"rows"} or \code{"cols"} to specify whether arrays 
are to be created from the rows or columns of the data frame. Defaults to
\code{"cols"}. If set to \code{"rows"}, the \code{df} must have row names.}

\item{global.}{A \code{boolean} indicating whether the variables
are to be loaded into the existing environment or returned as an object. 
Defaults to TRUE.}
}
\value{
NULL. Creates an R environment.
}
\description{
This function allows one to scrape rows and columns from an already
    existing \code{dataframe} to create an environment in R containing vectors
    corresponding to the rows or columns of the data frame or add such variables
    to an already-existing environment.
}
\examples{
\dontrun{
df <- data.frame(x = 1:3, y = 4:6)

dfEnv(df)
}

}
\keyword{df}
\keyword{environment,}
