% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonEnv.R
\name{jsonEnv}
\alias{jsonEnv}
\title{Create R Environment from JSON File}
\usage{
jsonEnv(json, global = TRUE)
}
\arguments{
\item{json}{A \code{string} to a \code{.json} file containing the environmental variables.}

\item{global.}{A \code{boolean} indicating whether the variables
are to be loaded into the existing environment or returned as an object. 
Defaults to TRUE.}
}
\value{
NULL. Creates an R environment.
}
\description{
This function allows one to scrape fields from a \code{.json} 
    file to create an environment in R containing the same key value pairs seen
    in the file.
}
\examples{
\dontrun{
jsonEnr("environment.json")
}

}
\keyword{environment,}
\keyword{json}
